--- contrib/ruby_event_store-rom/lib/ruby_event_store/rom/adapters/sql/changesets/create_events.rb
@@ -1,15 +1,0 @@
-# frozen_string_literal: true
-
-module RubyEventStore
-  module ROM
-    module SQL
-      module Changesets
-        class CreateEvents < ROM::Changesets::CreateEvents
-          def commit
-            relation.multi_insert(to_a)
-          end
-        end
-      end
-    end
-  end
-end
