+++ contrib/ruby_event_store-rom/lib/ruby_event_store/rom/changesets/create_stream_entries.rb
@@ -4,22 +4,14 @@ module RubyEventStore
   module ROM
     module Changesets
       class CreateStreamEntries < ::ROM::Changeset::Create
-        module Defaults
-          def self.included(base)
-            base.class_eval do
-              relation :stream_entries
+        relation :stream_entries
 
-              map do |tuple|
-                Hash(created_at: RubyEventStore::ROM::Types::DateTime.call(nil)).merge(tuple)
-              end
-              map do
-                map_value :created_at, ->(datetime) { datetime.to_time.localtime }
-              end
-            end
-          end
+        map do |tuple|
+          Hash(created_at: RubyEventStore::ROM::Types::DateTime.call(nil)).merge(tuple)
         end
-
-        include Defaults
+        map do
+          map_value :created_at, ->(datetime) { datetime.to_time.localtime }
+        end
       end
     end
   end
