+++ contrib/ruby_event_store-rom/lib/ruby_event_store/rom/changesets/update_events.rb
@@ -4,23 +4,43 @@ module RubyEventStore
   module ROM
     module Changesets
       class UpdateEvents < ::ROM::Changeset::Update
-        module Defaults
-          def self.included(base)
-            base.class_eval do
-              relation :events
+        relation :events
 
-              map(&:to_h)
-              map do
-                rename_keys timestamp: :created_at
-                map_value   :created_at, ->(time) { Time.iso8601(time).localtime }
-                map_value   :valid_at,   ->(time) { Time.iso8601(time).localtime }
-                accept_keys %i[event_id data metadata event_type created_at valid_at]
-              end
-            end
+        map(&:to_h)
+        map do
+          rename_keys timestamp: :created_at
+          map_value   :created_at, ->(time) { Time.iso8601(time).localtime }
+          map_value   :valid_at,   ->(time) { Time.iso8601(time).localtime }
+          accept_keys %i[event_id data metadata event_type created_at valid_at]
+        end
+
+        UPSERT_COLUMNS = %i[event_type data metadata valid_at].freeze
+
+        def commit
+          if SQL.supports_on_duplicate_key_update?(relation.dataset.db)
+            commit_on_duplicate_key_update
+          elsif SQL.supports_insert_conflict_update?(relation.dataset.db)
+            commit_insert_conflict_update
+          else
+            raise "Database doesn't support upserts: #{relation.dataset.db.adapter_scheme}"
           end
         end
 
-        include Defaults
+        private
+
+        def commit_on_duplicate_key_update
+          relation.dataset.on_duplicate_key_update(*UPSERT_COLUMNS).multi_insert(to_a)
+        end
+
+        def commit_insert_conflict_update
+          relation.dataset.insert_conflict(
+            # constraint: 'index_name',
+            target: :event_id,
+            update: UPSERT_COLUMNS.each_with_object({}) do |column, memo|
+              memo[column] = Sequel[:excluded][column]
+            end
+          ).multi_insert(to_a)
+        end
       end
     end
   end
